Updated how Course.class stores courses and updated AddCourses accordingly
To address the issue of having to update every student's academic history/timeline as well as course prerequisites every time an admin editted a course, firebase database was changed to store a course's unique ID, generated by UUID. this UUID was also stored as a field in Course.class.
To then update AddCourse accordingly, algorithm had to be made to be able to determine a course's ID given its course code.