Refactored how SharedPreferences stored a student's academic history
SharedPreferences used to store academic history as an ArrayList. The issue of which was that SharedPreferences didn't preserve the order of the arraylist, which caused major issues with using recycler view in conjunction with firebase realtime database. In order to maintain order of academic history, courses were now stored in firebase database as a string of format: "Course1ID;Course2ID;Course3IDl;...CoursenID"